logica de recompensa implementada

+++ src/main/java/br/com/upe/academia/AcademiaWeb/Entities/DTOs/AlunoDTOs.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 
 public class AlunoDTOs extends UsuarioDTOs{
 
-    private double saldoMoedas;
+    private int saldoMoedas;
 
     public AlunoDTOs(Aluno aluno) {
         super(aluno);
